+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add GvG tab (showing sector army details)
 * Fix GB calculation when self (still unsafe) is followed by owner
 * Fix (auto) battle info for newest version of FoE
+* Show images of buildings in plunder tab (thanks for contribution)
 
 # 0.3.2
 * Fix battle info for newest version of FoE